# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Ledger Polkadot Generic API
  description: This API provides a way to generate the required shortened metadata to sign a transaction on a Ledger device.
  version: 1.2.1
servers:
  - url: https://api.zondax.ch/dot/generic
    description: Production server
paths:
  /chains:
    get:
      summary: Get chains
      description: Get available chains info
      responses:
        '200':
          description: Chains info available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chains'
  /node/metadata/flush:
    post:
      summary: Flush cache
      description: Flush cache for a specific chain
      parameters:
        - name: chainId
          in: path
          required: true
          description: Ticket of the chain
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChainConfig'
      responses:
        '200':
          description: The cache was flushed successfully
        '404':
          description: Chain not found
  /node/metadata:
    post:
      summary: Retrieve complete node metadata
      description: Retrieve and cache complete metadata for a specific chain
      parameters:
        - name: chainId
          in: path
          required: true
          description: Ticket of the chain
          schema:
            type: string
      responses:
        '200':
          description: Chain metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChainConfig'
        '404':
          description: Chain not found
  /transaction/metadata:
    post:
      summary: Retrieve shortened metadata for a tx
      description: Retrieve shortened metadata, required to sign a tx on the ledger device
      parameters:
        - name: chainId
          in: path
          required: true
          description: Ticket of the chain
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TxPayload'
      responses:
        '200':
          description: Shortened metadata to sign tx
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortenedMetadata'
        '404':
          description: Chain not found
components:
  schemas:
    Metadata:
      type: object
      description: JSON metadata retrieved from the node
    ShortenedMetadata:
      type: object
      properties:
        shortenedMetadata:
          type: string
          description: Tx shortened metadata on hex format
    TxPayload:
      required:
        - callData
        - signedExtensions
        - chainConfig
      properties:
        callData:
          type: string
          description: Transaction payload to be signed, in serialized format
        signedExtensions:
          type: string
        chain:
          type: object
          properties:
            id:
              type: string
              description: ID of the chain to interact with
              enum:
                - ksm
                - ksm-hub
                - nodl
                - astr
                - dot
                - dot-hub
                - bnc

    ChainConfig:
      required:
        - id
      properties:
        id:
          type: string
          description: ID of the chain to interact with
          enum:
            - ksm
            - ksm-hub
            - nodl
            - astr
            - dot
            - dot-hub
            - bnc
    Chains:
      required:
        - id
        - name
        - url
      properties:
        id:
          type: string
          description: Chain ID
        name:
          type: string
          description: Chain name
        url:
          type: string
          description: Node API URL used internally to get metadata from